- Move NotificationApp to the base package as expected by Spring Boot conventions. - Move NotificationServiceErrorHandler to the notification.impl package, sitting along DefaultNotificationServiceImpl     - (not required for the error handler to function)
